userId数据类型的修改 Integer -> Long

+++ src/main/java/com/egovchina/partybuilding/partybuild/repository/TabSysUserMapper.java
@@ -178,7 +178,7 @@ public interface TabSysUserMapper {
      * @author FanYanGen
      * @date 2019/4/15 9:32
      **/
-    boolean checkIsExistByUserId(Integer userId);
+    boolean checkIsExistByUserId(Long userId);
 
     /**
      * desc: 根据身份证号码查询该用户是否存在
